+++ spec/dummy/app/models/author.rb
@@ -1,6 +1,6 @@
 class Author < ActiveRecord::Base
 
-  include Storext
+  include Storext.model
   store_attributes :data do
     name String
     title String, default: "Dr."