+++ spec/dummy/app/models/coffee.rb
@@ -1,6 +1,6 @@
 class Coffee < ActiveRecord::Base
 
-  include Storext.model
+  include Storext.model(data: {})
 
   store_attributes :data do
     name String