+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # TBA
 
 - Ensure that behaviour remains the same as `ActiveRecord::Base` when no attributes are defined
+- Allow options to be passed in with `include Storext.model`
 
 # 0.1.4